requirements.txt
Flask
pathlib
flask_sqlalchemy
